requirements-dev.txt
pyyaml>=5.4.1
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0
types-PyYAML>=6.0.0
